// core/apiClient/api.ts
export interface ScheduleDate {
    /**
     * 
     * @type {number}
     * @memberof ScheduleDate
     */
    year: number;
    /**
     * 
     * @type {number}
     * @memberof ScheduleDate
     */
    month: number;
    /**
     * 
     * @type {number}
     * @memberof ScheduleDate
     */
    date: number;
    /**
     * 
     * @type {number}
     * @memberof ScheduleDate
     */
    hour: number;
    /**
     * 
     * @type {number}
     * @memberof ScheduleDate
     */
    minute: number;
    /**
     * 
     * @type {number}
     * @memberof ScheduleDate
     */
    second: number;
}